+++ src/components/Dropdown/DropdownOption.tsx
@@ -4,12 +4,13 @@ import { st, classes } from './DropdownOption.st.css';
 import { TPAComponentProps } from '../../types';
 
 export interface DropdownOptionProps {
-  id: string;
-  value: string;
+  id?: string;
+  value?: string;
   subtitle?: string;
   isSelectable?: boolean;
   isSectionTitle?: boolean;
   icon?: React.ReactElement;
+  divider?: boolean;
 }
 
 export const DropdownOption: React.FC<DropdownOptionProps &