+++ src/components/Grid/Grid.tsx
@@ -1,11 +1,6 @@
 import * as React from 'react';
 import { st, classes, cssStates } from './Grid.st.css';
-import {
-  generateKey,
-  getGridStyle,
-  getMediaQueries,
-  itemsPerRowWidth,
-} from './GridUtils';
+import { getGridStyle, getMediaQueries, itemsPerRowWidth } from './GridUtils';
 import {
   TPAComponentsConsumer,
   TPAComponentsContext,
@@ -20,6 +15,7 @@ import {
 import { GridDataKeys, GridDataHooks } from './DataHooks';
 import { Item } from './Item/Item';
 import { TPAComponentProps } from '../../types';
+import { generateKey } from '../../common/random';
 
 export interface GridProps extends TPAComponentProps {
   children?: React.ReactNode[];