+++ src/components/Grid/GridUtils.spec.tsx
@@ -1,10 +1,5 @@
 import * as React from 'react';
-import {
-  itemsPerRowWidth,
-  getMediaQueries,
-  generateKey,
-  getGridStyle,
-} from './GridUtils';
+import { itemsPerRowWidth, getMediaQueries, getGridStyle } from './GridUtils';
 
 describe('GridUtils', () => {
   describe('itemsPerRowWidth', () => {
@@ -157,14 +152,6 @@ bottom: calc((${expectedRowGap}px / -2) - ${expectedDividerWidth});
           listWrapperClass: expectedListWrapperClass,
         }),
       ).toEqual(expectedStyle);
-    });
-  });
-
-  describe('generateKey', () => {
-    it('should add prefix to the id', () => {
-      const expectedPrefix = 'list_';
-
-      expect(generateKey(expectedPrefix)).toContain(expectedPrefix);
     });
   });
 });