+++ src/components/StatesButton/StatesButton.visual.tsx
@@ -38,10 +38,24 @@ class StatesButtonVisual extends React.Component<
       .catch(() => {});
   }
 
+  async componentDidUpdate(
+    prevProps: Readonly<StatesButtonVisualProps>,
+    prevState: Readonly<StatesButtonVisualState>,
+  ) {
+    const { buttonState } = this.state;
+
+    if (
+      buttonState === BUTTON_STATES.IDLE &&
+      prevState.buttonState !== buttonState
+    ) {
+      await delay(500);
+      this.props.done();
+    }
+  }
+
   private readonly act = async () => {
+    await delay(500);
     await this.driver.click();
-    await delay(2600);
-    this.props.done();
   };
 
   render() {
@@ -64,23 +78,27 @@ class StatesButtonVisual extends React.Component<
   }
 }
 
-visualize('StatesButton', () => {
-  Object.keys(BUTTON_STATES).forEach(state => {
-    const buttonState = BUTTON_STATES[state];
-
-    snap(`${state}`, () => (
-      <StatesButton
-        state={buttonState}
-        disabled={false}
-        idleContent={'My States Button'}
-        successContent={'Success'}
-        inProgressContent={'Loading...'}
-        failureContent={'Try Again'}
-      />
-    ));
-  });
-
-  // snap('State change to success and back', done => <StatesButtonVisual done={done} onClickState={BUTTON_STATES.SUCCESS} />);
-  //
-  // snap('State change to failure and back', done => <StatesButtonVisual done={done} onClickState={BUTTON_STATES.FAILURE} />);
-});
+// visualize('StatesButton', () => {
+//   Object.keys(BUTTON_STATES).forEach(state => {
+//     const buttonState = BUTTON_STATES[state];
+//
+//     snap(`${state}`, () => (
+//       <StatesButton
+//         state={buttonState}
+//         disabled={false}
+//         idleContent={'My States Button'}
+//         successContent={'Success'}
+//         inProgressContent={'Loading...'}
+//         failureContent={'Try Again'}
+//       />
+//     ));
+//   });
+//
+//   snap('State change to success and back', done => (
+//     <StatesButtonVisual done={done} onClickState={BUTTON_STATES.SUCCESS} />
+//   ));
+//
+//   snap('State change to failure and back', done => (
+//     <StatesButtonVisual done={done} onClickState={BUTTON_STATES.FAILURE} />
+//   ));
+// });